pyproject.toml
[project]
name = "mini-project-management-api"
version = "0.1.0"
description = "Mini Project Management System Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "sqlalchemy>=2.0.28",
    "asyncpg>=0.29.0",
    "pydantic[email]>=2.6.4",
    "pydantic-settings>=2.2.1",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "alembic>=1.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.5",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
